### 2025-04-03 #7985 @arfon
@nbb2 – thanks for your submission to JOSS. I'm going to start a scope query for the editorial team to judge whether this submission can be sent out for review. From a cursory inspection, I'm not sure this is the case.

The scope review will take a couple of weeks.
