### 2025-04-03 #7985 @nbb2
@arfon - Ok no problem. This is our first time submitting to the journal, so any feedback is welcome of course and we appriciate your time looking at the paper. Could you elaborate, if only briefly, what you mean by a "scope review" and what you think is out of scope from first glance?
