### 2025-04-23 #7985 @kyleniemeyer
@nbb2 the scope review was to make sure your submission was in an appropriate state for review (basically, we don't want to move things forward that are too small, not research software, or have some major issues that would immediately be raised in a review).

The editorial board has found that your submission in general meets our requirements for review, but there are some substantial issues that need to be resolved, so I'm going to pause this so you can address those.

These include:
- Some files that should not be committed to the repository or seem temporary (`.DS_Store`, `untitled2.m`).
- It appears there are some unnecessary re-implementations of basic functions, like the square root function. Can these be removed, or else documented why a custom reimplementation is needed?
- It seems the repository includes a third-party code authored by authors, `chebfun`. While I know that Matlab doesn't have the same packaging infrastructure as Python (for example) that allows easy specification of dependencies, something like a git submodule may be more appropriate, or otherwise including instructions that `chebfun` needs to be installed as a prerequisite.
- There also seems to be mixing of data and source files for tests, html folders in the `src` folder, etc.
- There is no documentation that I can see; the `docs` folder has some clearly incomplete files.

In particular, JOSS does require documentation and some sort of test suite for submissions.

If you think you can address these issues and get your submission to meet our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), I can pause this submission until you notify me that it's ready. Alternatively, if you think it will take too long, I can withdraw this, and you can resubmit to JOSS at a later time.
