### 2025-04-28 3fc5024
Delete src/my_cube.m
