+++ src/my_cube.m
@@ -1,5 +1,0 @@
-function solution = my_cube(x)
-
-    solution = x^3
-
-end
