+++ src/untitled2.m
@@ -1,8 +1,0 @@
-s = 10;
-H = zeros(s);
-
-for c = 1:s
-    for r = 1:s
-        H(r,c) = 1/(r+c-1);
-    end
-    endcc
