### 2025-04-28 d024f69
added VSS files
