+++ src/lj_124.m
@@ -1,7 +1,7 @@
-function y = my_126lj(eps_well,sig,r)
-%MY_LJ    Outputs Lennard Jones potential data.
-%   Y=MY_LJ(EPS_WELL,SIG,R) generates a float array containing
-%   a potential energy value for each r value using the 12-6 Lennard Jones
+function y = lj_124(eps_well,sig,r)
+%LJ_124    Outputs Lennard Jones potential data.
+%   Y=LJ_124(EPS_WELL,SIG,R) generates a float array containing
+%   a potential energy value for each r value using the 12-4 Lennard Jones
 %   potential. Units of potential energy are eV.
 %
 %   -- EPS_WELL must be the float LJ well depth.
@@ -10,5 +10,5 @@ function y = my_126lj(eps_well,sig,r)
 %   Angstrom.
 %
 %   See also RUN_CALCPOTENTIAL
-    y = 4*eps_well*(((sig./r).^12)-((sig./r).^6));
+    y = (0.5)*(3^1.5)*eps_well*(((sig./r).^12)-((sig./r).^4));
 end
