+++ src/lj_126.m
@@ -1,7 +1,7 @@
-function y = my_124lj(eps_well,sig,r)
-%MY_LJ    Outputs Lennard Jones potential data.
-%   Y=MY_124LJ(EPS_WELL,SIG,R) generates a float array containing
-%   a potential energy value for each r value using the 12-4 Lennard Jones
+function y = lj_126(eps_well,sig,r)
+%LJ_126    Outputs Lennard Jones potential data.
+%   Y=LJ_126(EPS_WELL,SIG,R) generates a float array containing
+%   a potential energy value for each r value using the 12-6 Lennard Jones
 %   potential. Units of potential energy are eV.
 %
 %   -- EPS_WELL must be the float LJ well depth.
@@ -10,5 +10,5 @@ function y = my_124lj(eps_well,sig,r)
 %   Angstrom.
 %
 %   See also RUN_CALCPOTENTIAL
-    y = (0.5)*(3^1.5)*eps_well*(((sig./r).^12)-((sig./r).^4));
+    y = 4*eps_well*(((sig./r).^12)-((sig./r).^6));
 end
