+++ src/rgtrun.m
@@ -1,7 +1,0 @@
-%read input deck
-
-%if potentiel calc == True
-
-%if potential fit ==True
-
-%if 

[…]
