### 2025-04-29 ab1e35b
added tab in 'calculate scattering integrals' tab to select chebfun folder, changed functions to work with this
