+++ test/PotentialFit/lj_fitoutputfile.m
@@ -1,12 +1,0 @@
-Potential_Type = "Lennard-Jones";
-minR = 1.000000;
-maxR = 10.000000;
-Tol = 1.000000e-12;
-eps_param = 2.000000e+00;
-sigma_param = 4.000000e+00;
-min_eps = 0.000000;
-max_eps = 10.000000;
-min_sigma = 0.000000;
-max_sigma = 10.000000;
-eps_start = 2.000000;
-sigma_start = 2.000000;

[…]
