### 2025-05-05 #7985 @JGranat8
@kyleniemeyer

We have addressed all of the issues you mentioned:
- Temporary files have been removed
- Basic functions have been removed
- Chebfun was removed from our src folder and the dependency is mentioned in the documentation.
- Test folder has been cleaned up and a separate testFiles folder was added.
- Documentation has been added in a new docs folder.

Please let us know if there are any other issues. Thanks so much for your help!
