### 2025-05-07 b36a7ad
fixed factor of dqrt(2) in VHS and VSS functions
