### 2025-07-21 7dde87c
fixed typo when using diffusion CS data as input for stopping CS array - name wasn't updated
