### 2025-10-20 cf6628c
Add standalone app version
