+++ rgtfun-standalone/paper.bib
@@ -1,0 +1,257 @@
+@article{amdur_experimental_1961,
+	title = {An experimental approach to the determination of gaseous transport properties at very high temperatures},
+	volume = {3},
+	issn = {00320633},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/0032063361902501},
+	doi = {10.1016/0032-0633(61)90250-1},
+	language = {en},
+	urldate = {2023-08-20},
+	journal = {Planetary and Space Science},
+	author = {Amdur},
+	month = feb,
+	year = {1961},
+	pages = {228--235},
+}
+@article{amdur_repulsive_1949,
+	title = {Repulsive {Interaction} {Potentials} at {Small} {Interaction} {Distances}: {He}–{He} and {H}–{H2} {Systems}},
+	volume = {17},
+	issn = {0021-9606, 1089-7690},
+	shorttitle = {Repulsive {Interaction} {Potentials} at {Small} {Interaction} {Distances}},
+	url = {https://pubs.aip.org/jcp/article/17/9/844/201134/Repulsive-Interaction-Potentials-at-Small},
+	doi = {10.1063/1.1747419},
+	language = {en},
+	number = {9},
+	urldate = {2023-08-20},
+	journal = {The Journal of Chemical Physics},
+	author = {Amdur, I.},
+	month = sep,
+	year = {1949},
+	pages = {844--845},
+}
+@article{amdur_scattering_1954,
+	title = {Scattering of {High}-{Velocity} {Neutral} {Particles}. {II}. {Helium}-{Helium}},
+	volume = {22},
+	issn = {0021-9606, 1089-7690},
+	url = {https://pubs.aip.org/jcp/article/22/4/664/204158/Scattering-of-High-Velocity-Neutral-Particles-II},
+	doi = {10.1063/1.1740144},
+	abstract = {Total collision cross sections have been measured for helium atoms, with energies between 500 and 2100 ev, scattered in room temperature helium. Measurements were made with two detectors of different geometric aperture, 13.4 and 0.96 minutes. In order to obtain meaningful potential energy information from such results, it is necessary to determine effective apertures by taking into account beam shape, intensity distribution, and for the case where the beam is wider than the detector, scattering to the detector from external regions of the beam. Potential energy functions have been derived from the cross sections obtained with the two detectors. The functions agree within a factor of about 1.3 as compared with 0.17 when geometric apertures are used directly. The results from both detectors have been represented by an average function, namely, V(r)=7.55×10−12/r5.94 ergsfor values of r between 1.27A and 1.59A.
+            The potential energy information obtained from the present experiments is in agreement with theoretical calculations of P. Rosen and H. Margenau within the uncertainty of such calculations. The information also appears to be consistent with that derived from measurements of gaseous compressibility and viscosity, for values of r between 1.90A and 2.30A.},
+	language = {en},
+	number = {4},
+	urldate = {2023-08-20},
+	journal = {The Journal of Chemical Physics},
+	author = {Amdur, I. and Harkness, A. L.},
+	month = apr,
+	year = {1954},
+	pages = {664--669},
+}
+@article{biersack_monte_1980,
+	title = {A {Monte} {Carlo} computer program for the transport of energetic ions in amorphous targets},
+	volume = {174},
+	copyright = {https://www.elsevier.com/tdm/userlicense/1.0/},
[…]
