### 2025-10-20 ca513ea
Delete rgtfun-standalone directory
