### 2025-10-20 ddd161e
Standalone app changes (clean branch)
