### 2026-01-30 #7985 @piyueh
Hi @bartlettnbb2 @nbb2, I’ve heard back from one of the potential reviewers, so we now only need one more. In the meantime, could you please restructure your paper to include the three dedicated sections described in [this comment](https://github.com/openjournals/joss-reviews/issues/7985#issuecomment-3761312660)?
