### 2026-02-03 9c924d4
Restructured paper to satisfy new JOSS requirements
