### 2026-02-03 6c9ea7c
Restructured paper to satisfy new JOSS requirements
