+++ paper.md
@@ -54,11 +54,10 @@ $VHS$ - variablele hard sphere - the collision rule used for the DSMC
 # Statement of Need
 RGTFun is a MATLAB app designed for the efficient calculation of scattering integrals and transport coefficients of elastic collisions. While the solution for classical scattering, as well as its application for determining gas transport coefficients, has been known for over approximately a century [@chapman_mathematical_1990], the numerical codes are usually kept as closed-source or are printed in textbooks in older programming languages such as FORTRAN [@maitland_intermolecular_1987]. To our knowledge, existing publicly available tools (e.g., SRIM for stopping power calculations) offer limited support for user-adjustable interatomic potentials (especially across energy regimes) in end-to-end workflows that compute scattering integrals and the resulting transport coefficients.
 
-RGTFun is open source and implemented in MATLAB, leveraging its numerical computing capabilities and App Designer for the graphical interface. We welcome issues and pull requests, and we encourage users to extend the app to their applications. Furthermore, the code features a well-designed graphical user interface (GUI) to facilitate the step by step process of going from an intermolecular potential to macroscopic transport coefficients such as viscosity and diffusion coefficients. The ethos of the project is to decrease the learning curve of going from quantum chemistry calculations of intermolecular potential energy surfaces to usable transport coefficients in CFD or PIC codes.  
+RGTFun is open source and implemented in MATLAB, leveraging its numerical computing capabilities and App Designer for the graphical interface. Furthermore, the code features a well-designed graphical user interface (GUI) to facilitate the step by step process of going from an intermolecular potential to macroscopic transport coefficients such as viscosity and diffusion coefficients. The ethos of the project is to decrease the learning curve of going from quantum chemistry calculations of intermolecular potential energy surfaces to usable transport coefficients in CFD or PIC codes.  
 
-The paper briefly reviews the theory behind bimolecular scattering and how cross sections of elastic scattering events are used to calculate macroscopic transport properties. Then, four validation cases of the code are reviewed. RGTFun is validated across four levels of fidelity: (1) a single-collision analytic benchmark by comparing Coulomb scattering angles against the exact classical solution; (2) a standard screened-potential benchmark by comparing ZBL scattering angles to the Biersack “Magic Formula” [@biersack_monte_1980]; (3) a cross-tool benchmark by comparing ZBL-based nuclear stopping cross sections to SRIM results for H–H [@ziegler_treatise_1985] [@ziegler_srim_2010]; and (4) an end-to-end macroscopic benchmark by comparing predicted argon viscosity and self-diffusion coefficients to NIST reference data [@kestin_equilibrium_1984].
-
-This app has already been incorporated into the authors' computational workflow and has been used by our collaborators at ASML to generate coefficients for downstream modeling. 
+# Research Impact Statement
+RGTFun has been incorporated into the authors’ computational workflow and has been used by collaborators at ASML to generate coefficients for downstream modeling. By providing an end-to-end workflow from an intermolecular potential to macroscopic transport coefficients (e.g., viscosity and diffusion), RGTFun lowers the barrier for translating potential energy surfaces from quantum chemistry into inputs for CFD or PIC simulations. In addition to interactive GUI use, the modular scripts support batch execution for parameter sweeps and integration into existing pipelines. The validation cases in this paper provide a reference analysis spanning analytic scattering, screened-potential benchmarks, cross-tool comparison to SRIM, and comparison to NIST transport data.
 
 # Software Design 
[…]
