### 2026-02-09 #7985 @piyueh
@nbb2 I’ve added the first reviewer. The second reviewer has accepted the invitation, and I’m waiting for his GitHub handle to add him here. Once he is added and your paper is revised according to the new JOSS requirements, we can proceed with the review.

@Stephen-Armstrong Thank you for accepting the invitation. This is the pre-review issue, where editors perform initial checks and coordinate with authors to identify reviewers. Once we add the second reviewer, a formal review issue will be created, and you will proceed with the review there.
