### 2026-02-09 #7985 @piyueh
Sure. Please let me know when it's ready. If you want to check the compilation of your `paper.md`, feel free to ask the bot to do it (via `@editorialbot generate pdf`)
