### 2026-02-09 #7985 @nbb2
Is it possible to delete that article proof from the history of generated proofs? There is some content I would not like to be included in the history. Sorry for the inconvience.
