### 2026-02-09 #7985 @piyueh
@drobnyjt Thanks for accepting the invitation. This is the pre-review issue, where editors do the initial checks and coordinate with the authors to find reviewers. We asked the authors to update their paper to reflect a recent policy change at JOSS. Once the updated version is in, we’ll move ahead with the formal review process. (We'll open a new GitHub issue for reviewers and authors to communicate for the formal review process.)
