### 2026-02-09 #7985 @piyueh
@nbb2 Is the current version the one you don't want? Or you have already overwrote it and it's still in the history? Though the generated PDF is also version controlled by git (here: https://github.com/openjournals/joss-papers/tree/joss.07985/joss.07985), I think the history is force-pushed and overwritten every time a new PDF is generated, as I currently only see one commit there. But I'll check with the tech support. In the mean time, you can generate a new PDF and see if the history is overwritten.
