### 2026-02-10 6b0858f
Restructured paper to satisfy new JOSS requirements
