+++ paper.md
@@ -47,16 +47,29 @@ $N_a$ - Avogadro's number (atoms / mol)
 $P_{vap}$ - equilibrium vapor pressure (Pa)  
 $k_B$ - Boltzmann's constant ($J \cdot K^{-1}$)  
 $T$ - temperature (K)  
-$M_{Sn}$ - mass of a tin atom (Kg)  
 $d_{12}$ - collision diameter used in the VHS collision rule for DSMC simulations  
 $VHS$ - variablele hard sphere - the collision rule used for the DSMC
 
 # Statement of Need
-RGTFun is a MATLAB app designed for the efficient calculation of scattering integrals and transport coefficients of elastic collisions. While the solution for classical scattering, as well as its application for determining gas transport coefficients, has been known for over approximately a century [@chapman_mathematical_1990], the numerical codes are usually kept as closed source codes or are printed in textbooks in older programming languages such as FORTRAN [@maitland_intermolecular_1987].
+RGTFun is a MATLAB app designed for the efficient calculation of scattering integrals and transport coefficients of elastic collisions. While the solution for classical scattering, as well as its application for determining gas transport coefficients, has been known for over approximately a century [@chapman_mathematical_1990], the numerical codes are usually kept as closed-source or are printed in textbooks in older programming languages such as FORTRAN [@maitland_intermolecular_1987]. To our knowledge, existing publicly available tools (e.g., MagnumPI for calculation of scattering angles and cross sections [@magnumpi]; SRIM for stopping power calculations [@ziegler_srim_2010]) offer limited support for user-adjustable interatomic potentials (especially across energy regimes) in end-to-end workflows that compute scattering integrals and the resulting transport coefficients.
 
-This new code is open source and uses the well known dynamically compiled language MATLAB. Furthermore, the code features a well designed graphical user interface (GUI) to facilitate the step by step process of going from an intermolecular potential to macroscopic transport coefficients such as viscosity and diffusion coefficient. The ethos of the project is to decrease the learning curve of going from quantum chemistry calculations of intermolecular potential energy surfaces to usable transport coefficients in CFD or PIC codes.   
+The closest existing open-source tool to RGTFun is MagnumPI, which focuses on computing scattering angles and (differential) cross sections. In contrast, RGTFun extends the workflow beyond cross sections to compute macroscopic transport coefficients (e.g., viscosity and diffusion). Because MagnumPI does not provide transport coefficients, it also does not support downstream parameterizations that are fitted to transport behavior (e.g., variable hard sphere coefficients), whereas RGTFun includes these end-to-end capabilities.
 
-The paper briefly reviews the theory behind bimolecular scattering and how cross sections of elastic scattering events are used to calculate macroscopic transport properties. Then, four validation cases of the code are reviewed.
+RGTFun is open source and implemented in MATLAB, leveraging its numerical computing capabilities and App Designer for the graphical interface. Furthermore, the code features a well-designed graphical user interface (GUI) to facilitate the step by step process of going from an intermolecular potential to macroscopic transport coefficients such as viscosity and diffusion coefficients. The ethos of the project is to decrease the learning curve of going from quantum chemistry calculations of intermolecular potential energy surfaces to usable transport coefficients in CFD or PIC codes.  
+
+# Research Impact Statement
[…]
