+++ paper.bib
@@ -157,6 +157,13 @@
 	year={1931},
 	publisher={IOP Publishing}
 }
+@misc{magnumpi,
+  title        = {MagnumPI},
+  author       = {Jesper Janssen},
+  year         = {2018},
+  howpublished = {\url{https://gitlab.com/magnumpi/magnumpi}},
+  note         = {GitLab repository, accessed 2026-02-09}
+}
 @book{maitland_intermolecular_1987,
 	address = {Oxford},
 	edition = {Repr},
