+++ paper.md
@@ -55,7 +55,7 @@ RGTFun is a MATLAB app designed for the efficient calculation of scattering inte
 
 The closest existing open-source tool to RGTFun is MagnumPI, which focuses on computing scattering angles, cross sections, and collision integrals for several (but not all) intermolecular potentials. In contrast, RGTFun extends the workflow beyond this to compute macroscopic transport coefficients (e.g., viscosity and diffusion) as well as simplified collision model parameters for the variable hard sphere (VHS) and variable soft sphere (VSS) models. RGTfun aims to provide an end-to-end calculation with a helpful graphical user interface (GUI). Furthermore, RGTfun shows the user the results of intermediate calculations (such as the distance of closest approach) so the user can verify that the algorithm is behaving correctly at each step. 
 
-RGTFun is open source and implemented in MATLAB, leveraging its numerical computing capabilities and App Designer for the graphical interface. Furthermore, the code features a well-designed graphical user interface (GUI) to facilitate the step by step process of going from an intermolecular potential to macroscopic transport coefficients such as viscosity and diffusion coefficients. The ethos of the project is to decrease the learning curve of going from quantum chemistry calculations of intermolecular potential energy surfaces to usable transport coefficients in CFD or PIC codes.  
+RGTFun is open-source and implemented in MATLAB, leveraging its numerical computing capabilities and App Designer for the graphical interface. Furthermore, the code features a well-designed graphical user interface (GUI) to facilitate the step by step process of going from an intermolecular potential to macroscopic transport coefficients such as viscosity and diffusion coefficients. The ethos of the project is to decrease the learning curve of going from quantum chemistry calculations of intermolecular potential energy surfaces to usable transport coefficients in CFD or PIC codes.  
 
 # Research Impact Statement
 RGTFun has been incorporated into the authors’ computational workflow to generate coefficients for downstream modeling. By providing an end-to-end workflow from an intermolecular potential to macroscopic transport coefficients (e.g., viscosity and diffusion), RGTFun lowers the barrier for translating potential energy surfaces from quantum chemistry into inputs for CFD or PIC simulations. In addition to interactive GUI use, the modular scripts support batch execution for parameter sweeps and integration into existing pipelines. The validation cases in this paper provide a reference analysis spanning analytic scattering, screened-potential benchmarks, cross-tool comparison to SRIM, and comparison to NIST transport data.
