### 2026-02-10 #7985 @JGranat8
Hi @piyueh, we are done revising the paper to fulfill the new JOSS requirements. Thanks for your patience!

-Jack
