### 2026-02-10 #7985 @piyueh
@nbb2 @JGranat8 Thanks for the revision. Could you please also include a dedicated **“State of the Field”** section, as this is part of the new policy? Otherwise, the pre-check will not pass (see the check result [here](https://github.com/openjournals/joss-reviews/issues/7985#issuecomment-3879342161)). Please refer to the updated policy [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain). I believe you already have similar content in your **“Statement of Need”** section, so you can likely separate that material into its own section.

I apologize for any inconvenience this may cause, and thank you for your understanding.
