### 2026-02-10 f23467b
Added State of the Field section
