+++ paper.md
@@ -53,9 +53,12 @@ $VHS$ - variablele hard sphere - the collision rule used for the DSMC
 # Statement of Need
 RGTFun is a MATLAB app designed for the efficient calculation of scattering integrals and transport coefficients of elastic collisions. While the solution for classical scattering, as well as its application for determining gas transport coefficients, has been known for over approximately a century [@chapman_mathematical_1990], the numerical codes are usually kept as closed-source or are printed in textbooks in older programming languages such as FORTRAN [@maitland_intermolecular_1987]. To our knowledge, existing publicly available tools (e.g., MagnumPI for calculation of scattering angles and cross sections [@magnumpi]; SRIM for stopping power calculations [@ziegler_srim_2010]) offer limited support for user-adjustable interatomic potentials (especially across energy regimes) in end-to-end workflows that compute scattering integrals and the resulting transport coefficients.
 
-The closest existing open-source tool to RGTFun is MagnumPI, which focuses on computing scattering angles, cross sections, and collision integrals for several (but not all) intermolecular potentials. In contrast, RGTFun extends the workflow beyond this to compute macroscopic transport coefficients (e.g., viscosity and diffusion) as well as simplified collision model parameters for the variable hard sphere (VHS) and variable soft sphere (VSS) models. RGTfun aims to provide an end-to-end calculation with a helpful graphical user interface (GUI). Furthermore, RGTfun shows the user the results of intermediate calculations (such as the distance of closest approach) so the user can verify that the algorithm is behaving correctly at each step. 
-
 RGTFun is open-source and implemented in MATLAB, leveraging its numerical computing capabilities and App Designer for the graphical interface. Furthermore, the code features a well-designed graphical user interface (GUI) to facilitate the step by step process of going from an intermolecular potential to macroscopic transport coefficients such as viscosity and diffusion coefficients. The ethos of the project is to decrease the learning curve of going from quantum chemistry calculations of intermolecular potential energy surfaces to usable transport coefficients in CFD or PIC codes.  
+
+# State of the Field
+The closest existing open-source tool to RGTFun is MagnumPI, which focuses on computing scattering angles, cross sections, and collision integrals for several (but not all) intermolecular potentials [@magnumpi]. SRIM is widely used for stopping power calculations, but it is not designed to calculate transport coefficients and does not support user-adjustable interatomic potentials in the manner required here [@ziegler_srim_2010].
+
+We chose to develop RGTFun rather than contribute to existing tools because our target use case requires capabilities beyond collision-scale quantities: RGTFun computes macroscopic transport coefficients (e.g., viscosity and diffusion) and also provides simplified collision-model parameters for variable hard sphere (VHS) and variable soft sphere (VSS) models. In addition, RGTFun is organized as an end-to-end, user-guided workflow with intermediate diagnostics (e.g., distance of closest approach) to help users verify expected behavior at each stage. This combination of user-adjustable potentials across energy regimes, end-to-end transport outputs, and a validation-oriented GUI fills a gap not addressed by existing open-source tools.
 
 # Research Impact Statement
 RGTFun has been incorporated into the authors’ computational workflow to generate coefficients for downstream modeling. By providing an end-to-end workflow from an intermolecular potential to macroscopic transport coefficients (e.g., viscosity and diffusion), RGTFun lowers the barrier for translating potential energy surfaces from quantum chemistry into inputs for CFD or PIC simulations. In addition to interactive GUI use, the modular scripts support batch execution for parameter sweeps and integration into existing pipelines. The validation cases in this paper provide a reference analysis spanning analytic scattering, screened-potential benchmarks, cross-tool comparison to SRIM, and comparison to NIST transport data.
