### 2026-02-10 #7985 @JGranat8
Hi @piyueh, I've now revised the paper to include a dedicated "State of the Field" section.

-Jack
