### 2026-02-10 #9948 @piyueh
@Stephen-Armstrong, @drobnyjt Thank you for agreeing to review this submission.

If this is your first time reviewing for JOSS, please begin by asking the bot to generate your checklist:
`@editorialbot generate my checklist`

We encourage an interactive review process. Please feel free to comment here or open an issue in the authors’ software repository whenever you need clarification or would like to request changes, rather than waiting until you have completed a full review pass. The author (@nbb2) will respond promptly and engage throughout the process.

Additional resources to support your review (in addition to the information provided in the top comment of this thread):

1. Review criteria: https://joss.readthedocs.io/en/latest/review_criteria.html
2. Checklist item explanations: https://joss.readthedocs.io/en/latest/review_checklist.html
