### 2026-04-15 2314f6f
Last part of fix for #28: reduced mass used as input for test did not match what was used for expected result. This is fixed now and all tests in the test suite are running correctly.
