### 2026-04-16 81e5f3a
fixed typos in journal text, addresses issue #29
